added gather_test to large_vector and added comments in large array and vector tests

+++ tests/nightly/test_large_array.py
@@ -1505,9 +1505,11 @@ def test_pad():
 
 def test_gather():
     arr = mx.nd.ones((LARGE_X, SMALL_Y))
-    idx = mx.nd.random.randint(0, LARGE_X, SMALL_X, dtype=np.int64)
+    idx = mx.nd.random.randint(0, LARGE_X, SMALL_X)
+    # Calls gather_nd internally
     tmp = arr[idx]
     assert np.sum(tmp[0] == 1) == SMALL_Y
+    # Calls gather_nd internally
     arr[idx] += 1
     assert np.sum(arr[idx[0]] == 2) == SMALL_Y
